Update LLM.py
The section for separately detecting emotions has been deleted, and the program directly generates emotional words in the original text, and then uses the pattern matching function to directly find related emotions in the answers. If there is one, it is set to answer_feeling. If not, it is set to "smile", which is an improvement. twice the speed